Identifying Which Line We are "On" Based on Normal Vector Orientation
In findNextIntersection, I was able to orient the normal vector that I passed in based on the rotation of the given rhomb at the intersection, called the orientationDirection. Then, I added the intersection point vector and the orientationDirection vector to get a translated point, which would live on either of the two gridline indices. From there, I had trouble using the containPoint method, as it was being finnicky and not returning the correct answers though the logic was sound. Also, the intersectLine method returns an intersection point that is “incorrect” where it is scaled by a factor of -1. This causes an issue, but it seems that the codebase is working in a sense around this error. Trying to change it to the correct formula, the containsPoint method causes a ‘Bad Intersection’ error to be thrown – very weird – it seems that method is incredibly sensitive.

The way in which Line objects are created is indeed correct. So I wrote a new method called lineContainsTranslatedPoint, which takes in a Line object and a 2D vector and outputs a Boolean. We calculate LHS = a*x + b*y and then RHS = -c (accounts for the weird -1 factor on our intersection point that we have to deal with) . Then, it takes abs(LHS-RHS) < small EPSILON. I first just equated LHS ==RHS, but I got the wrong Booleans returned due to miniscule differences in RHS & LHS due to computational round-off error.

In a case structure, I passed in our lines using defineLine with each one of our gridline indices and then the intersection point; Via debugging output, as adjacent tile goes around each side, it is able to identify each line we are supposed to move along based on it finding the translatedPoint on that line via the normal vector; Everything looks correct

Now we know how to classify which GLI as the one we are “on” and then the one that is just “crossing”, which begins the stages for iterating through the other 4 familes, finding the closest intersections for each one; Then, from those, sorting the one out with least distance to find the next intersection we move to to generate a tile